+++ CHANGES.md
@@ -4,6 +4,7 @@
 
 ### Changed
 
+* Add Item and Collection `PATCH` endpoints with support for [RFC 6902](https://tools.ietf.org/html/rfc6902) and [RFC 7396](https://tools.ietf.org/html/rfc7386) in the `TransactionExtension`
 - remove support of `cql-json` in Filter extension ([#840](https://github.com/stac-utils/stac-fastapi/pull/840))
 
 ### Fixed
@@ -175,7 +176,6 @@
 ## [3.0.0] - 2024-07-29
 
 Full changelog: https://stac-utils.github.io/stac-fastapi/migrations/v3.0.0/#changelog
-
 **Changes since 3.0.0b3:**
 
 ### Changed